pyproject.toml
[project]
name = "dbsamizdapper"
version = "0.1.0"
description = "Blissfully naive PostgreSQL database object manager for views, materialized views, functions, and triggers"
authors = [{name = "Josh Brooks", email = "josh@catalpa.io"}]
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE.txt"}
keywords = ["postgresql", "database", "views", "materialized-views", "django", "sql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Database",
]
dependencies = [
    "pytest-cov>=7.0.0",
    "toposort>=1.10",
]

[project.urls]
Homepage = "https://github.com/catalpainternational/dbsamizdapper"
Documentation = "https://github.com/catalpainternational/dbsamizdapper#readme"
Repository = "https://github.com/catalpainternational/dbsamizdapper"
Issues = "https://github.com/catalpainternational/dbsamizdapper/issues"

[project.optional-dependencies]
django = [
    "django>=4.2,<5.0",
    "django-stubs>=4.2.0",
]

psycopg3 = [
    "psycopg[binary]>=3.1.9",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.3.0",
    "types-toposort>=1.10.0.0",
    "python-dotenv>=1.0.0",
]

testing = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "psycopg2-binary",
    "types-psycopg2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dbsamizdat"]

[tool.hatch.build.targets.sdist]
exclude = [
    "tests/",
    "sample_app/",
]

[tool.pylint.format]
max-line-length = "119"

[tool.ruff]
line-length = 119
target-version = "py312"
exclude = [
    ".venv",
    "__pycache__",
    ".git",
    ".pytest_cache",
    ".mypy_cache",
    "dist",
    "build",
    "*.egg-info",
    "pnds_db/metabase/*",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "ARG001", # unused function argument (common in protocol implementations)
    "UP038", # isinstance with tuple (not always compatible with X | Y syntax)
    "B904",  # raise without from (intentional in some exception handling cases)
    "ARG002", # unused method args (common with pytest fixtures)
    "ARG005", # unused lambda args (sometimes needed for API compatibility)
    "B024",  # abstract base class without abstract methods (protocol pattern)
    "E402",  # module level import not at top (examples may have comments)
    "SIM117", # nested with statements (sometimes clearer)
    "B017",  # pytest.raises(Exception) (acceptable in some test cases)
]

[tool.ruff.lint.isort]
known-first-party = ["dbsamizdat"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[[tool.mypy.overrides]]
module = [
    "psycopg",
    "psycopg2"
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["dbsamizdat"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/sample_app/*",
    "*/examples/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
skip_covered = false
precision = 2

[tool.coverage.html]
directory = "htmlcov"
